kafka_data_producer/requirements.txt
APScheduler==3.2.0
cassandra-driver==3.7.1
click==6.6
confluent-kafka==0.9.2
Flask==0.11.1
funcsigs==1.0.2
futures==3.0.5
googlefinance==0.7
itsdangerous==0.24
Jinja2==2.8
kafka-python==1.3.1
MarkupSafe==0.23
pytz==2016.7
schedule==0.3.2
setuptools==29.0.1
simplejson==3.10.0
six==1.10.0
tzlocal==1.3
Werkzeug==0.11.11
wheel==0.30.0a0
yahoo-finance==1.4.0